requirements.txt
apscheduler
fastapi
pytz
python-telegram-bot
python-telegram-bot[job-queue]
requests
motor
gunicorn
uvicorn